Bump to 50ad171.
remove workaround for a crash with color emojis on some systems, now fixed in libXft 2.3.5

https://gitlab.freedesktop.org/xorg/lib/libxft/-/blob/libXft-2.3.5/NEWS

bump version to 6.4

Ref.
https://git.suckless.org/dwm/commit/970f37697358574e127019eb0ee2f5725ec05ce0.html
https://git.suckless.org/dwm/commit/50ad171eea9db5ccb36fce2592e047c3282975ff.html

Consequently the color emoji patch has been inverted into no color emoji, keeping the workaround in the code base for those that are on systems with older versions of the Xft library.