Fix and improve BAR_WINTITLEACTIONS_PATCH (#442)
Introduced an unhideall function that can be used to reveal all hidden client in the current view.

Added a change to unmanage to avoid changing the window state from IconicState to WithdrawnState when the window manager stops managing windows (which happens when restarting).

Previously a hidden window would disappear when restarting dwm because:
   - when hidden the window would be unmapped and having the window state of IconicState
   - when restarting the window would remain unmapped while the window state would become WithdrawnState
   - which would prevent the window from being picked up by the scan function when the window manager starts